Move splice out of the forecast stage
